# dev_core/api/openapi.yaml
openapi: 3.0.3
info:
  title: InduFrame Control Plane API
  version: 0.1.0
  description: dev_core 面向 dev_ide 的控制面契约。
servers:
  - url: /api/v1
...
paths:
...
  /logs:
    get:
      operationId: listAuditLogs
      tags: [auditLog]
      summary: listAuditLogs
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        default:
          description: 失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
    delete:
      operationId: deleteAuditLogs
      tags: [auditLog]
      summary: deleteAuditLogs
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        default:
          description: 失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /logs/{logId}:
    get:
      operationId: getAuditLog
      tags: [auditLog]
      summary: getAuditLog
      parameters:
        - name: logId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        default:
          description: 失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /logs/export:
    get:
      operationId: exportAuditLogs
      tags: [auditLog]
      summary: exportAuditLogs
      responses:
        '200':
          description: 成功
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        default:
          description: 失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /logs/stats:
    get:
      operationId: getAuditLogStats
      tags: [auditLog]
      summary: getAuditLogStats
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        default:
          description: 失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /logs/recent-activities:
    get:
      operationId: listRecentActivities
      tags: [auditLog]
      summary: listRecentActivities
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        default:
          description: 失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    agentTokenAuth:
      type: http
      scheme: bearer
      bearerFormat: NodeAgentToken
  schemas:
...
    ApiResponse:
      type: object
      required: [code, msg, data, reqId]
      properties:
        code:
          type: integer
        msg:
          type: string
        data:
          nullable: true
        reqId:
          type: string
security:
  - bearerAuth: []